refactor: doc.style.get_fmt()
If the filetype defines a template for a section, the template is used
as is, it should take care of the control character in the template it
defines.

This is because we don't know what a custom formatter can return, while
we know that the built-in templates need the control character to be
added at generation time. Example:

The basic template for 'params' is:

	{char}param %s: ...

{char} must be fetched at generation time, so it can't be in the
template, that isn't a funcref. It must be added later. But we can do
this for built-in formatters, because we know that the control character
hasn't been added yet, but if the custom formatter doesn't want the
control character? Hence it must be accepted as-is.